requirements.txt
streamlit
scipy.stats

